master delete

+++ masterSync/masterDelete.js
@@ -6,6 +6,8 @@ export default async function masterDelete(body) {
 
   const contactToDelete = (await client.api(`${process.env.MASTER_PATH}`).get()).value.find(directory => directory.spouseName === body.id);
 
+  if (!contactToDelete) return;
+
   const res = await client.api(`${process.env.DIRECTORY_PATH}/${contactToDelete.id}`).delete();
   console.log(res)
 }
